+++ ChangeLog.md
@@ -1,5 +1,11 @@
 ﻿# Rhetos release notes
 
+## 5.6.0 (2026-02-03)
+
+* New generic property filters: `ContainsAny`, `StartsWithAny`, `EndsWithAny`.
+  Each matches records where the property value contains, starts with, or ends with any string from a provided list.
+  Also available in [REST API](https://github.com/rhetos/restgenerator#filters).
+
 ## 5.5.0 (2025-04-16)
 
 ### New features